Revert "Merge pull request #2086 from scipp/transform-coord-transpose"
This reverts commit ceeb5e367bc3e1985fa8d62410cd85585cce2553, reversing
changes made to 6c11b77ed5a94fa168f78b48ab3ba419a00a56d0.
