+++ python/src/scipp/coords.py
@@ -6,7 +6,6 @@ import inspect
 import warnings
 from typing import Union, List, Dict, Tuple, Callable
 from . import Variable, DataArray, Dataset, bins, VariableError
-from .utils.dimensions import find_bin_edges
 
 
 def _argnames(func):
@@ -72,28 +71,6 @@ def _produce_coord(obj, name):
     return obj.coords[name]
 
 
-def _move_to_back(lis, val):
-    lis.remove(val)
-    lis.append(val)
-    return lis
-
-
-def _move_bin_edge_to_innermost(var, edge_dim):
-    return var.transpose(_move_to_back(var.dims, edge_dim))
-
-
-def _store_coord(obj, name, coord):
-    edges = find_bin_edges(coord, obj)
-    if not edges:
-        obj.coords[name] = coord
-    elif len(edges) == 1:
-        obj.coords[name] = _move_bin_edge_to_innermost(coord, edges[0])
-    else:
-        raise NotImplementedError(
-            'Coordinates with more than one bin-edge dimension are not supported. '
-            f'Got coord {coord} with bin edges {edges}.')
-
-
 class CoordTransform:
     Graph = Dict[Union[str, Tuple[str, ...]], Union[str, Callable]]
 
@@ -155,7 +132,7 @@ class CoordTransform:
             out = {name: out}
         self._rename.setdefault(dim, []).extend(out.keys())
         for key, coord in out.items():
-            _store_coord(self.obj, key, coord)
+            self.obj.coords[key] = coord
         if self.obj.bins is not None:
             if isinstance(out_bins, Variable):
                 out_bins = {name: out_bins}
