+++ python/src/scipp/html/formatting_html.py
@@ -9,7 +9,6 @@ from html import escape
 
 from .._scipp import core as sc
 from .resources import load_icons
-from ..utils.dimensions import find_bin_edges
 
 BIN_EDGE_LABEL = "[bin-edge]"
 VARIANCE_PREFIX = "σ² = "
@@ -169,6 +168,23 @@ def summarize_coord(dim, var, ds=None):
 
 def summarize_mask(dim, var, ds=None):
     return summarize_variable(str(dim), var, is_index=False, embedded_in=ds)
+
+
+def find_bin_edges(var, ds):
+    """
+    Checks if the coordinate contains bin-edges.
+    """
+    bin_edges = []
+    for idx, dim in enumerate(var.dims):
+        length = var.shape[idx]
+        if not ds.dims:
+            # Have a scalar slice.
+            # Cannot match dims, just assume length 2 attributes are bin-edge
+            if length == 2:
+                bin_edges.append(dim)
+        elif dim in ds.dims and ds.shape[ds.dims.index(dim)] + 1 == length:
+            bin_edges.append(dim)
+    return bin_edges
 
 
 def summarize_coords(coords, ds=None):
